+++ packages/waggle/src/libp2p/websocketOverTor.ts
@@ -21,6 +21,7 @@ class Discovery extends EventEmitter {
     this.tag = 'channel_18'
   }
 
+  stop() {}
   start() {}
   end() {}
 }
@@ -44,7 +45,11 @@ class WebsocketsOverTor extends WebSockets {
     let socket
     let maConn
     try {
-      socket = await this._connect(ma, { websocket: this._websocketOpts, ...options, localAddr: this.localAddress })
+      socket = await this._connect(ma, {
+        websocket: this._websocketOpts,
+        ...options,
+        localAddr: this.localAddress
+      })
     } catch (e) {
       log.error('error connecting to %s. Details: %s', ma, e.message)
       throw e
@@ -67,13 +72,13 @@ class WebsocketsOverTor extends WebSockets {
     }
   }
 
-  async _connect (ma, options: any = {}) {
-    if (options.signal && options.signal.aborted) {
+  async _connect(ma: multiaddr, options: any = {}) {
+    if (options.signal?.aborted) {
       throw new AbortError()
     }
     const cOpts = ma.toOptions()
     log('dialing %s:%s', cOpts.host, cOpts.port)
-    const myUri = `${toUri(ma)}/?remoteAddress=${encodeURIComponent(this.localAddress)}`
+    const myUri = `${toUri(ma) as string}/?remoteAddress=${encodeURIComponent(this.localAddress)}`
     const rawSocket = connect(myUri, Object.assign({ binary: true }, options))
     if (!options.signal) {
       await rawSocket.connected()
@@ -84,6 +89,7 @@ class WebsocketsOverTor extends WebSockets {
 
     // Allow abort via signal during connect
     let onAbort
+    // eslint-disable-next-line
     const abort = new Promise((resolve, reject) => {
       onAbort = () => {
         reject(new AbortError())
@@ -112,6 +118,7 @@ class WebsocketsOverTor extends WebSockets {
     }
     const server = createServer(options, async (stream, request) => {
       let maConn, conn
+      // eslint-disable-next-line
       const query = url.parse(request.url, true).query
       log('query', query.remoteAddress)
       try {
@@ -125,7 +132,7 @@ class WebsocketsOverTor extends WebSockets {
         conn = await upgrader.upgradeInbound(maConn)
       } catch (err) {
         log.error('inbound connection failed to upgrade', err)
-        return maConn && maConn.close()
+        return maConn?.close()
       }
 
       log('inbound connection %s upgraded', maConn.remoteAddr)
@@ -151,7 +158,7 @@ class WebsocketsOverTor extends WebSockets {
       return server.close()
     }
 
-    listener.listen = (ma) => {
+    listener.listen = (ma: multiaddr) => {
       listeningMultiaddr = ma
 
       return server.listen(ma.toOptions())
@@ -165,12 +172,13 @@ class WebsocketsOverTor extends WebSockets {
         throw new Error('Listener is not ready yet')
       }
 
-      const ipfsId = listeningMultiaddr.getPeerId()
+      const ipfsId: string = listeningMultiaddr.getPeerId()
 
       // Because TCP will only return the IPv6 version
       // we need to capture from the passed multiaddr
       if (listeningMultiaddr.toString().indexOf('ip4') !== -1) {
         let m = listeningMultiaddr.decapsulate('tcp')
+        // eslint-disable-next-line
         m = m.encapsulate('/tcp/' + address.port + '/ws')
         if (listeningMultiaddr.getPeerId()) {
           m = m.encapsulate('/p2p/' + ipfsId)
@@ -178,8 +186,8 @@ class WebsocketsOverTor extends WebSockets {
 
         if (m.toString().indexOf('0.0.0.0') !== -1) {
           const netInterfaces = os.networkInterfaces()
-          Object.keys(netInterfaces).forEach((niKey) => {
-            netInterfaces[niKey].forEach((ni) => {
+          Object.keys(netInterfaces).forEach(niKey => {
+            netInterfaces[niKey].forEach(ni => {
               if (ni.family === 'IPv4') {
                 multiaddrs.push(multiaddr(m.toString().replace('0.0.0.0', ni.address)))
               }
@@ -193,15 +201,6 @@ class WebsocketsOverTor extends WebSockets {
       return multiaddrs
     }
     return listener
-  }
-
-  createListener (options = {}, handler) {
-    if (typeof options === 'function') {
-      handler = options
-      options = {}
-    }
-
-    return this.prepareListener({ handler, upgrader: this._upgrader }, options)
   }
 }