+++ packages/waggle/src/socket/DataServer.ts
@@ -1,5 +1,6 @@
 import express from 'express'
 import { createServer, Server } from 'http'
+// eslint-disable-next-line
 const socketio = require('socket.io')
 
 export class DataServer {
@@ -18,9 +19,19 @@ export class DataServer {
     this.io = socketio(this.server)
   }
 
-  public listen = (): void => {
-    this.server.listen(this.PORT, () => {
-      console.debug(`Server running on port ${this.PORT}`)
+  public listen = async (): Promise<void> => {
+    return await new Promise((resolve) => {
+      this.server.listen(this.PORT, () => {
+        console.debug(`Server running on port ${this.PORT}`)
+        resolve()
+      })
+    })
+  }
+
+  public close = async (): Promise<void> => {
+    return await new Promise((resolve) => {
+      this.server.close()
+      resolve()
     })
   }
 }