+++ packages/waggle/src/socket/events/message.ts
@@ -5,6 +5,5 @@ export const message = (socket, message) => {
 }
 
 export const directMessage = (socket, message) => {
-  console.log('emitting direct message to zbaylite')
   socket.emit(EventTypesServer.DIRECT_MESSAGE, message)
 }