+++ packages/waggle/src/storage/storage.ts
@@ -6,7 +6,10 @@ import OrbitDB from 'orbit-db'
 import KeyValueStore from 'orbit-db-kvstore'
 import EventStore from 'orbit-db-eventstore'
 import PeerId from 'peer-id'
-import { message as socketMessage, directMessage as socketDirectMessage, directMessage } from '../socket/events/message'
+import {
+  message as socketMessage,
+  directMessage as socketDirectMessage
+} from '../socket/events/message'
 import { loadAllMessages, loadAllDirectMessages } from '../socket/events/allMessages'
 import { EventTypesResponse } from '../socket/constantsReponse'
 import fs from 'fs'
@@ -73,10 +76,10 @@ export class Storage {
   private publicChannelsEventsAttached: boolean = false
 
   public async init(libp2p: any, peerID: PeerId): Promise<void> {
-    log('STORAGE: Entered init')
     const ipfsRepoPath = path.join(this.zbayDir, 'ZbayChannels')
     const orbitDbDir = path.join(this.zbayDir, 'OrbitDB')
     createPaths([ipfsRepoPath, orbitDbDir])
+
     this.ipfs = await IPFS.create({
       libp2p: () => libp2p,
       preload: { enabled: false },
@@ -99,7 +102,11 @@ export class Storage {
     log('5/6')
     await this.initAllConversations()
     log('6/6')
-    log('STORAGE: Finished init')
+  }
+
+  public async stopOrbitDb() {
+    await this.orbitdb.stop()
+    await this.ipfs.stop()
   }
 
   public async loadInitChannels() {
@@ -121,7 +128,7 @@ export class Storage {
     })
 
     this.channels.events.on('replicated', () => {
-    log('REPLICATED: CHANNELS')
+      log('REPLICATED: CHANNELS')
     })
     await this.channels.load({ fetchEntryTimeout: 15000 })
     log('ALL CHANNELS COUNT:', Object.keys(this.channels.all).length)
@@ -135,12 +142,16 @@ export class Storage {
         write: ['*']
       }
     })
-    this.messageThreads.events.on('replicated', async () => {
-      await this.messageThreads.load({ fetchEntryTimeout: 2000 })
-      const payload = this.messageThreads.all
-      this.io.emit(EventTypesResponse.RESPONSE_GET_PRIVATE_CONVERSATIONS, payload)
-      this.initAllConversations()
-    })
+    this.messageThreads.events.on(
+      'replicated',
+      // eslint-disable-next-line
+      async () => {
+        await this.messageThreads.load({ fetchEntryTimeout: 2000 })
+        const payload = this.messageThreads.all
+        this.io.emit(EventTypesResponse.RESPONSE_GET_PRIVATE_CONVERSATIONS, payload)
+        await this.initAllConversations()
+      }
+    )
     await this.messageThreads.load({ fetchEntryTimeout: 2000 })
     log('ALL MESSAGE THREADS COUNT:', Object.keys(this.messageThreads.all).length)
   }
@@ -152,38 +163,46 @@ export class Storage {
       }
     })
 
-    this.directMessagesUsers.events.on('replicated', async () => {
-      await this.directMessagesUsers.load({ fetchEntryTimeout: 2000 })
-      const payload = this.directMessagesUsers.all
-      this.io.emit(EventTypesResponse.RESPONSE_GET_AVAILABLE_USERS, payload)
-      log('REPLICATED USERS')
-    })
+    this.directMessagesUsers.events.on(
+      'replicated',
+      // eslint-disable-next-line
+      async () => {
+        await this.directMessagesUsers.load({ fetchEntryTimeout: 2000 })
+        // await this.directMessagesUsers.close()
+        const payload = this.directMessagesUsers.all
+        this.io.emit(EventTypesResponse.RESPONSE_GET_AVAILABLE_USERS, payload)
+        log('REPLICATED USERS')
+      }
+    )
     try {
       await this.directMessagesUsers.load({ fetchEntryTimeout: 2000 })
     } catch (err) {
       log.error(err)
     }
     log('ALL USERS COUNT:', Object.keys(this.directMessagesUsers.all).length)
-    // console.log('ALL USERS COUNT:', Object.keys(this.directMessagesUsers.all))
   }
 
   async initAllChannels() {
     console.time('initAllChannels')
-    await Promise.all(Object.values(this.channels.all).map(async channel => {
-      if (!this.publicChannelsRepos.has(channel.address)) {
-        await this.createChannel(channel.address, channel)
-      }
-    }))
+    await Promise.all(
+      Object.values(this.channels.all).map(async channel => {
+        if (!this.publicChannelsRepos.has(channel.address)) {
+          await this.createChannel(channel.address, channel)
+        }
+      })
+    )
     console.timeEnd('initAllChannels')
   }
 
   async initAllConversations() {
     console.time('initAllConversations')
-    await Promise.all(Object.keys(this.messageThreads.all).map(async conversation => {
-      if (!this.directMessagesRepos.has(conversation)) {
-        await this.createDirectMessageThread(conversation)
-      }
-    }))
+    await Promise.all(
+      Object.keys(this.messageThreads.all).map(async conversation => {
+        if (!this.directMessagesRepos.has(conversation)) {
+          await this.createDirectMessageThread(conversation)
+        }
+      })
+    )
     console.timeEnd('initAllConversations')
   }
 
@@ -236,7 +255,10 @@ export class Storage {
     loadAllMessages(this.io, this.getAllChannelMessages(db), channelAddress)
   }
 
-  public async subscribeForChannel(channelAddress: string, channelInfo?: IChannelInfo): Promise<void> {
+  public async subscribeForChannel(
+    channelAddress: string,
+    channelInfo?: IChannelInfo
+  ): Promise<void> {
     let db: EventStore<IMessage>
     let repo = this.publicChannelsRepos.get(channelAddress)
 
@@ -245,7 +267,7 @@ export class Storage {
     } else {
       db = await this.createChannel(channelAddress, channelInfo)
       if (!db) {
-        console.log(`Can't subscribe to channel ${channelAddress}`)
+        log(`Can't subscribe to channel ${channelAddress}`)
         return
       }
       repo = this.publicChannelsRepos.get(channelAddress)
@@ -282,7 +304,7 @@ export class Storage {
     channelData?: IChannelInfo
   ): Promise<EventStore<IMessage>> {
     if (!channelAddress) {
-      console.log('No channel address, can\'t create channel')
+      log("No channel address, can't create channel")
       return
     }
     log('BEFORE CREATING NEW ZBAY CHANNEL')
@@ -314,7 +336,6 @@ export class Storage {
     await this.directMessagesUsers.load({ fetchEntryTimeout: 2000 })
     const payload = this.directMessagesUsers.all
     this.io.emit(EventTypesResponse.RESPONSE_GET_AVAILABLE_USERS, payload)
-
   }
 
   public async initializeConversation(address: string, encryptedPhrase: string): Promise<void> {
@@ -329,14 +350,16 @@ export class Storage {
 
     this.directMessagesRepos.set(address, { db, eventsAttached: false })
     await this.messageThreads.put(address, encryptedPhrase)
-    this.subscribeForDirectMessageThread(address)
+    await this.subscribeForDirectMessageThread(address)
   }
 
   public async subscribeForAllConversations(conversations) {
     console.time('subscribeForAllConversations')
-    await Promise.all(conversations.map(async channel => {
+    await Promise.all(
+      conversations.map(async channel => {
         await this.subscribeForDirectMessageThread(channel)
-    }))
+      })
+    )
     console.timeEnd('subscribeForAllConversations')
   }
 
@@ -349,7 +372,7 @@ export class Storage {
     } else {
       db = await this.createDirectMessageThread(channelAddress)
       if (!db) {
-        console.log(`Can't subscribe to direct messages thread ${channelAddress}`)
+        log(`Can't subscribe to direct messages thread ${channelAddress as string}`)
         return
       }
       repo = this.directMessagesRepos.get(channelAddress)
@@ -375,11 +398,9 @@ export class Storage {
     }
   }
 
-  private async createDirectMessageThread(
-    channelAddress: string
-  ): Promise<EventStore<IMessage>> {
+  private async createDirectMessageThread(channelAddress: string): Promise<EventStore<IMessage>> {
     if (!channelAddress) {
-      console.log('No channel address, can\'t create channel')
+      log("No channel address, can't create channel")
       return
     }
 
@@ -402,8 +423,6 @@ export class Storage {
     return db
   }
 
-
-
   public async sendDirectMessage(channelAddress: string, message) {
     await this.subscribeForDirectMessageThread(channelAddress) // Is it necessary? Yes it is atm
     log('STORAGE: sendDirectMessage entered')
@@ -419,7 +438,6 @@ export class Storage {
     log('STORAGE: getAvailableUsers entered')
     await this.directMessagesUsers.load({ fetchEntryTimeout: 2000 })
     const payload = this.directMessagesUsers.all
-    log(`STORAGE: getAvailableUsers ${payload}`)
     this.io.emit(EventTypesResponse.RESPONSE_GET_AVAILABLE_USERS, payload)
     log('emitted')
   }