+++ packages/waggle/src/utils.ts
@@ -1,4 +1,5 @@
 import fs from 'fs'
+import fp from 'find-free-port'
 
 export function createPaths(paths: string[]) {
   for (const path of paths) {
@@ -9,6 +10,24 @@ export function createPaths(paths: string[]) {
 }
 
 export function fetchAbsolute(fetch: Function): Function {
-  return (baseUrl: string) => (url, ...otherParams) =>
+  return (baseUrl: string) => (url: string, ...otherParams) =>
     url.startsWith('/') ? fetch(baseUrl + url, ...otherParams) : fetch(url, ...otherParams)
+}
+
+export const getPorts = async (): Promise<{
+  socksPort: number
+  libp2pHiddenService: number
+  controlPort: number
+  dataServer: number
+}> => {
+  const [controlPort] = await fp(9151)
+  const [socksPort] = await fp(9052)
+  const [libp2pHiddenService] = await fp(7788)
+  const [dataServer] = await fp(4677)
+  return {
+    socksPort,
+    libp2pHiddenService,
+    controlPort,
+    dataServer
+  }
 }